+++ sub-packages/bionemo-esm2/src/bionemo/esm2/model/finetune/datamodule.py
@@ -24,9 +24,11 @@ from nemo.utils import logging
 
 from bionemo.core.data.multi_epoch_dataset import IdentityMultiEpochDatasetWrapper, MultiEpochDatasetResampler
 from bionemo.esm2.data import tokenizer
-from bionemo.esm2.model.finetune.dataset import InMemoryCSVDataset
-from bionemo.esm2.model.finetune.finetune_regressor import InMemorySingleValueDataset
-from bionemo.esm2.model.finetune.finetune_token_classifier import InMemoryPerTokenValueDataset
+from bionemo.esm2.model.finetune.dataset import (
+    InMemoryCSVDataset,
+    InMemoryPerTokenValueDataset,
+    InMemorySingleValueDataset,
+)
 from bionemo.llm.data import collate
 from bionemo.llm.data.datamodule import MegatronDataModule
 from bionemo.llm.utils.datamodule_utils import infer_num_samples
