+++ sub-packages/bionemo-esm2/src/bionemo/esm2/model/finetune/dataset.py
@@ -25,7 +25,16 @@ from torch import Tensor
 from torch.utils.data import Dataset
 
 from bionemo.esm2.data import tokenizer
+from bionemo.llm.data.collate import MLM_LOSS_IGNORE_INDEX
+from bionemo.llm.data.label2id_tokenizer import Label2IDTokenizer
 from bionemo.llm.data.types import BertSample
+
+
+__all__: Sequence[str] = (
+    "InMemoryCSVDataset",
+    "InMemorySingleValueDataset",
+    "InMemoryPerTokenValueDataset",
+)
 
 
 class InMemoryCSVDataset(Dataset):
@@ -123,3 +132,88 @@ class InMemoryCSVDataset(Dataset):
             transformed_label
         """
         return label
+
+
+class InMemorySingleValueDataset(InMemoryCSVDataset):
+    """An in-memory dataset that tokenizes strings into BertSample instances."""
+
+    def __init__(
+        self,
+        data_path: str | os.PathLike,
+        tokenizer: tokenizer.BioNeMoESMTokenizer = tokenizer.get_tokenizer(),
+        seed: int = np.random.SeedSequence().entropy,  # type: ignore
+    ):
+        """Initializes a dataset for single-value regression fine-tuning.
+
+        This is an in-memory dataset that does not apply masking to the sequence.
+
+        Args:
+            data_path (str | os.PathLike): A path to the CSV file containing sequences and labels (Optional)
+            tokenizer (tokenizer.BioNeMoESMTokenizer, optional): The tokenizer to use. Defaults to tokenizer.get_tokenizer().
+            seed: Random seed for reproducibility. This seed is mixed with the index of the sample to retrieve to ensure
+                that __getitem__ is deterministic, but can be random across different runs. If None, a random seed is
+                generated.
+        """
+        super().__init__(data_path=data_path, tokenizer=tokenizer, seed=seed)
+
+    def transform_label(self, label: float) -> Tensor:
+        """Transform the regression label.
+
+        Args:
+            label: regression value
+
+        Returns:
+            tokenized label
+        """
+        return torch.tensor([label], dtype=torch.float)
+
+
+class InMemoryPerTokenValueDataset(InMemoryCSVDataset):
+    """An in-memory dataset of labeled strings, which are tokenized on demand."""
+
+    def __init__(
+        self,
+        data_path: str | os.PathLike,
+        tokenizer: tokenizer.BioNeMoESMTokenizer = tokenizer.get_tokenizer(),
+        seed: int = np.random.SeedSequence().entropy,  # type: ignore
+    ):
+        """Initializes a dataset for per-token classification fine-tuning.
+
+        This is an in-memory dataset that does not apply masking to the sequence.
+
+        Args:
+            data_path (str | os.PathLike): A path to the CSV file containing sequences and labels (Optional)
+            tokenizer (tokenizer.BioNeMoESMTokenizer, optional): The tokenizer to use. Defaults to tokenizer.get_tokenizer().
+            seed: Random seed for reproducibility. This seed is mixed with the index of the sample to retrieve to ensure
+                that __getitem__ is deterministic, but can be random across different runs. If None, a random seed is
+                generated.
+        """
+        super().__init__(data_path=data_path, tokenizer=tokenizer, seed=seed)
+        label_tokenizer = Label2IDTokenizer()
+        self.label_tokenizer = label_tokenizer.build_vocab("CHE")
+        self.label_cls_eos_id = MLM_LOSS_IGNORE_INDEX
+
+    def transform_label(self, label: str) -> Tensor:
+        """Transform the sequence label by tokenizing them.
+
+        This method tokenizes the secondary structure token sequences.
+
+        Args:
+            label: secondary structure token sequences to be transformed
+
+        Returns:
+            tokenized label
+        """
+        label_ids = torch.tensor(self.label_tokenizer.text_to_ids(label))
+
+        # # for multi-label classification with BCEWithLogitsLoss
+        # tokenized_labels = torch.nn.functional.one_hot(label_ids, num_classes=self.label_tokenizer.vocab_size)
+        # cls_eos = torch.full((1, self.label_tokenizer.vocab_size), self.label_cls_eos_id, dtype=tokenized_labels.dtype)
+
+        # for multi-class (mutually exclusive) classification with CrossEntropyLoss
+        tokenized_labels = label_ids
+        cls_eos = torch.tensor([self.label_cls_eos_id], dtype=tokenized_labels.dtype)
+
+        # add cls / eos label ids with padding value -100 to have the same shape as tokenized_sequence
+        labels = torch.cat((cls_eos, tokenized_labels, cls_eos))
+        return labels
